+++ src/api/resources/internalOrganizations/types/index.ts
@@ -1,6 +1,4 @@
 export * from "./DeleteResourceData.js";
 export * from "./DeleteResourceResponse.js";
 export * from "./EnrolledReward.js";
-export * from "./MerchantNetwork.js";
-export * from "./MerchantNetworkName.js";
 export * from "./OrganizationPaginationMetadata.js";